Remove sphinx pinning (#338)
Now that we've moved our docs publishing away from read the docs to our
own hosting we are no longer constrained by the build environment for
read the docs as we're running builds in github actions now. This means
we no longer need to pin our sphinx version, or jupyter-sphinx version
for our docs jobs. This commit updates the docs requirements list and
the sphinx config to no longer pin these requirements.