Add section to retworkx for networkx users guide on matrix converters
This commit adds documentation to the retworkx for networkx users guide
on the differences between networkx and retworkx when it comes to matrix
generation. Right now retworkx only has an `adjacency_matrix()` function
which is equivalent to networkx's `to_numpy_array()` and returns a numpy
array of the adjacency matrix. This is different from networkx's
`adjacency_matrix()` function which returns a scipy sparse array (and is
just an alias for networkx's `to_scipy_sparse_matrix()`). Documenting
this gap is important because it's easy to get confused by these API
differences.

This difference is fundamentally because we can effeciently create numpy
arrays in retworkx by just pass a pointer to numpy with a 2d array
that was constructed natively in rust (using rust-numpy) while other
matrix formats (such as scipy sparse matrices) don't expose a C api so
to create those objects we'd have to go through Python. Eventually we
might expand the API to include pure python functions to do additional
conversions (although I worry about increasing the external dependency
list) but numpy arrays will still be the default because of the inherent
performance benefits.

Fixes #342