Prepare 0.9.0 release (#339)
* Prepare 0.9.0 release

This commit starts the preparation for the 0.9.0 release. It moves all
the release notes for the release into a separate folder (to distinguish
them from notes for future releases) and also reworks the content of the
release notes and updates the documentation for the release. This commit
should be the last one merged before releasing 0.9.0 to ensure we've
moved and updated all the release notes before the release.

Partially implements #328

* Move new release notes

* Add readme section on optional dependencies

* Fix lint

* Move new release notes

* Update readme based on review comments

* Update index intro based on review comments

* Update release prelude to mention pydot_draw

* Move new release notes

* Use mpl_draw in docstrings

* Add example to graphviz_draw doc

* Update add-transitivity-function-c32d2bbbf3857c40.yaml

* Apply suggestions from code review

Co-authored-by: Ivan Carvalho <ivan.ivancps.cn@gmail.com>

* Update prepare-0.9-dab64f0197fc6233.yaml

* Update pydot-drawer-c5df0aa830679748.yaml

* Update README.md

* Update multigraph release note

* Move and update last new release note

* Add examples to spring layout release note

* Add examples to random layout

* Update add-random-layout-c1c2751be971e5d0.yaml