Bump version post-release
Now that 0.9.0 is out the door this commit bumps the package version on
the main development branch to enable landing new features. It doesn't
look like we're going to have a bugfix release (and if we do we can just
branch from the 0.9.0 tag). Bumping the version strings enables us t
differentiate the development branch from the previous release.