Fix execution of graphivz tests in CI
This commit fixes the tox job configuration so we actually install the
requirements for running the graphviz drawer tests. To run the graphviz
drawer tests you need graphviz installed in addition to the retworkx
extras requirements 'graphviz' (which contains pillow and pydot).
However, prior to #339 the extra was called 'pydot' not 'graphviz' and
this was never updated in the tox.ini. This has resulted in the graphivz
drawer tests all being skipped because we're not installing the python
requirements in the test environment. This commit fixes this oversight
and configures tox to install the correct extra so we can run the tests
(at least on linux where we install graphviz in CI).