Add vf2 mapping functions
This commit adds a new function vf2_mapping which is used to get the
isomorphic node mapping between 2 graphs. It works in the same way as
is_isomorphic() but instead of simply returning a boolean whether the
graphs are isomorphic it returns the node id mapping from first to
second for the matching.