Remove unnecessary borrows from code
In the current nightly version of rust's clippy a new check for finding
unnecessary borrows on variables was added. This check has identified
several places we're passing a borrowed references that wasn't
necessary. This commit just fixes those places identified by nightly
clippy so that when rust 1.55 is released we don't have any issues.