Handle non-existent files in edgelist cleanup
In the test_invalid_return_type_weight_fn edge list test methods we're
occasionally seeing CI failures caused by an error in cleanup. The test
method calls a write_edge_list() method of the graph objects that will
raise an exception, however to prevent leaking a temp file an addCleanup
call was added protectively. However, whether the file gets written or
not is not deterministic on every platform (apparently especially on
macOS). To avoid a spurious CI failure this commit adds a custom
function to handle any exceptions raised during cleanup and treat them
as non-fatal.