Really fix macOS CI failures
In #380 we updated the tests to fix some tests when in macOS CI the
,write_edge_list() negative tests were not actually writing a file to
disk which was causing a failure in the cleanup phase. However, that PR
missed a test case which had the same bug and the issue isn't actually
fixed. This commit corrects that oversight and handles a non-existent
file in cleanup for the test cases which were missed.