Fix incorrect result with dijkstra path functions (#388)
This commit fixes an issue with the dijkstra path functions where they
were returning the incorrect result in the case of multiple paths
between nodes in the graph. Previously, in cases of parallel paths the
inner rust dijkstra function was not updating the path if it had a
shorter weight than the already found path. This could result in an
incorrect shortest path being returned, depsite the shortest length
being correct. This fixes this by making sure we always update the path
when we set the distance. Additionally, after fixing this several of the
dijkstra path tests began to fail because they were not returning the
correct result.

Fixes #387