Add dag_weighted_longest_path_length() function
As a follow up to #383 what the underlying Qiskit user case for #381
would like is to use floats for the edge weighting. However for the
existing dag_longest_path_length() function our hands are tied with the
return type being an unsigned integer for backwards compatibility. To
workaround that limitation this commit adds a new function
dag_weighted_longest_path_length() which requires a weight function and
works with f64 weights. To accomplish this, the inner longest_path() rust
function is made generic for any numeric type returned by the callback
function, which required leveraging the num-traits crate. The new
pyfunction is then just a thin wrapper that works with f64 instead of
usize.