Improve negative test coverage
The coverage report showed that we had no test coverage of the cycle
error when the function is called on a graph with a cycle. This commit
corrects it. The other coverage hole was of the None case of a max_by()
call which always will be defined. The None case was just raising with a
more legible but still nondescript error message, "Encountered something
unexpected". So instead of building that manually and wasting LOC this
just uses unwrap which will raise a panic exception that is less friendly
but at least provides debug information if somehow that path got
executed without defining any elements (it returns in the empty case
well beforehand).