Add weighted longest path functions (#393)
* Add dag_weighted_longest_path_length() function

As a follow up to #383 what the underlying Qiskit user case for #381
would like is to use floats for the edge weighting. However for the
existing dag_longest_path_length() function our hands are tied with the
return type being an unsigned integer for backwards compatibility. To
workaround that limitation this commit adds a new function
dag_weighted_longest_path_length() which requires a weight function and
works with f64 weights. To accomplish this, the inner longest_path() rust
function is made generic for any numeric type returned by the callback
function, which required leveraging the num-traits crate. The new
pyfunction is then just a thin wrapper that works with f64 instead of
usize.

* Raise a ValueError if NaN is returned as a weight

A float NaN value returned by the callback can't be used with the
longest path function so we should just raise when it's returned instead
of potentially panicing at runtime if a NaN is returned.

* Improve negative test coverage

The coverage report showed that we had no test coverage of the cycle
error when the function is called on a graph with a cycle. This commit
corrects it. The other coverage hole was of the None case of a max_by()
call which always will be defined. The None case was just raising with a
more legible but still nondescript error message, "Encountered something
unexpected". So instead of building that manually and wasting LOC this
just uses unwrap which will raise a panic exception that is less friendly
but at least provides debug information if somehow that path got
executed without defining any elements (it returns in the empty case
well beforehand).

* Add dag_weighted_longest_path() too

This commit adds another new function dag_weighted_longest_path() which
is the dual of dag_longest_path() but using floats and a mandatory
weight_fn. Since all the hard work is already done adding the weighted
float variant for dag_weighted_longest_path_length() it makes sense to
just add both together.

* Documentation fixes from code review

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>