Remove unnecessary .collect() from _floyd_warshall() (#394)
This commit removes an unnecessary .collect() on the node indices from
being used in _floyd_warshall(). To build the output container the
function was looping over the node indices. Previously it was building a
Vec by iterating over the contents of graph.node_indices() and then
proceeding to iterate over the vec. Instead we can remove the
intermediate Vec and directly iterate over the NodeIndices iterator
returned by graph.node_indices().

This was caught by clippy on rust nightly which will save us from having
to fix this at a future date when a new rust version is released.