Build non-x86 architecture wheels in QEMU emulation (#398)
* Build non-x86 architecture wheels in QEMU emulation

Since the introduction of tight travis quotas for open source projects
we have not been running travis CI on retworkx for ~2 months. Since we
can't rely on travis to reliably run our non-x86 testing and release
builds we need to come up with an alternative solution. However, travis
is the only avaiable CI provider with non-x86 nodes available for free
to open source projects (which is why we were using them, this quota
issue has been a long time coming). To workaround this commit tries
to leverage QEMU emulation in github actions for the release wheel
builds. In practice this is at least an order of magnitude slower than
running on a VM but the github actions job time limit is 6 hrs which
hopefully should be enough time.

* DNM: test wheel builds

* Only run non-x86 on linux

* Only run ppc64le in the ppc64le job

* Attempt to set per python cross build variable

* Revert "DNM: test wheel builds"

It looks like all the wheel builds work now. The s390x might have issues
with timeouts as the emulation seems slower than other platforms, but at
least for right now it's working. (we can just manually spin up builds
if we hit time outs at release.

This reverts commit dc38c42d1fe17ffdfcc04d92e0ef055c3d461fb8.