Docs CI fixes
This commit makes 2 changes to the docs ci configuration. The first
fixes a bash syntax issue in the deploy script which was blocking the
publishing of stable versions of the docs on release. With that typo
fixed we won't manually need to reupload documentation to the stable
path on release. The second fix is a future facing fix for reno, since
we switched to using main as the default branch if we ever introduced a
stable branch to retworkx a bug in reno<3.4.0 would require a master
branch to exist. This commit bumps the minimum reno version to 3.4.0 and
configures it to use main as the default branch in case we ever start
using stable branches.