Add metric closure function (#390)
* Add metric closure function

This commit adds a metric closure function to retworkx. The metric
closure of a graph is the complete graph in which each edge is weighted
by the shortest path distance between the nodes in the graph. This is
the first step towards implementing the minimum steiner graph function
because the first step for that is to compute the metric closure for the
graph.

Additionally, this function is entirely self contained to a new rust module
steiner_tree.rs (which we'll eventually add the steiner_tree() function
too) in the interest of showing the eventual organizational structure we
want for #300 (although having an algorithms namespace might make sense
for #300). In a separate PR addressing #300 we should start to
reorganize the rust code in lib.rs like this to make things easier to
find.

Partially implements #389

* Move docs to Other Algorithms section

* Fix allocation size for output vector

* Rework disconnected graph logic

The previous check for a disconnected graph only caught graphs with
completely disconnected nodes and didn't detect all disconnected
graphs. This commit fixes the logic by check the first node in the graph
has a path to all other nodes.

* Fix handling of empty graph to not overflow

* Update src/lib.rs

Co-authored-by: Ivan Carvalho <ivan.ivancps.cn@gmail.com>

* Rework final loop to be over hashmap instead of indices

Co-authored-by: Ivan Carvalho <ivan.ivancps.cn@gmail.com>