Switch to iterating over node indices in metric_closure
In #390 one of the last changes made to the PR before merging was:
https://github.com/Qiskit/retworkx/pull/390/commits/3353b4b4cc3a7f859812145e25eb6290dc446ec4
which changed the final loop constructing the metric closure edges from
iterating over node indices and pulling the path from the hashmap for
that index to just iterating over the hashmap. That however had the
unintended side effect of introducing non-determinism to the output as
the iteration order over a hashmap isn't guaranteed. This was causing
failures in the steiner tree function as the metric closure edge order
can affect the computed tree. This commit fixes this by switching back
to using the node id order for the final output generation and adds a
comment as to why.