Handle graph edge cases
This commit adds handling for 3 edge cases with the input graph, it's
disconnected, it's empty, or it's a single node and returns an
appropriate early return (NaN, 0, inf respectively). To handle the
disconnected case a new helper function is_connected() is added to the
connectivity module which does a quick bfs traversal from a node in the
graph to see if all other nodes are reachable or not.