Add option to set null value for matrix generators
This commit adds a new kwarg option to the *distance_matrix() and
*adjacency_matrix() functions, null_value, which is used to adjust the
float value that's used to represent the absence of an edge or a path in
the graph. This was previously hard coded as always 0.0 but this enables
users to specify an alternate value instead, such as NaN or infinity.