Add test case with equal distances
This adds a test case with an input graph that has equal distances. It
is a test based on a review comment [1] that was previously producing an
incorrect output (that wasn't even a tree). After fixing the underlying
issue it would be good to have this tested to ensure we don't regress on
it in the future.

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

[1] https://github.com/Qiskit/retworkx/pull/392#discussion_r687646066