Fix handling of null value in from_adjacency_matrix (#409)
* Fix handling of null value in from_adjacency_matrix

This commit fixes an issue with the handling of null values in the
from_adjacency_matrix for PyDiGraph and PyGraph. Previously, the
function incorrectly treated any value <= 0 as a null value and didn't
add an edge for that. But the behavior should have been != instead of
less than or equal to. Additionally, the null value of 0.0 was hard
coded into the constructor methods which precluded using different
values. This commit fixes both issues by adding a new kwarg, null_value,
that lets users specify the null value they'd like to use. Additionally,
the behavior is changed to just use a not equals comparison to enable
using negative values (assuming null_value is zero).

Fixes #408

* Fix NaN handling

* Fix release note typo