Return an iterator over all valid Vf2 mappings (#376)
Building on top of #368, this PR refactors Vf2 algorithm and implements a vf2_mapping function
that returns a custom python class that acts as a python iterator that lazily evaluates all valid Vf2
mappings. The "trick" that allows the lazily evaluation is to take ownership of a copy of the graphs.
We can do this (with minimal overhead) since we were already creating new graphs in case of
id_order=False or nodes_removed=True. At the same time, it adjusts Vf2ppSorter to handle
graphs with non-contiguous node ids.

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>

* Add vf2 mapping functions

This commit adds a new function vf2_mapping which is used to get the
isomorphic node mapping between 2 graphs. It works in the same way as
is_isomorphic() but instead of simply returning a boolean whether the
graphs are isomorphic it returns the node id mapping from first to
second for the matching.

* Fix reverse mapping construction with vf2pp mapping

This commit fixes the construction of the reverse map when id_order=False. The previous
versions of this did not correctly build the map in all cases.

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

* Add vf2pp remapping failure test case

* Don't remap indices if no match found

* Fix clippy failures

* Use NodeMap for return type

* Run cargo fmt

* return an iterator over all valid vf2 mappings

* update release note

* lint

* fix Vf2ppSorter to use node id instead of edge insertion order as a tie breaker

* fix clippy warning

* update text signature

* update tests since now we can deterministically predict the output mapping

* add call_limit kwarg in Vf2 algorithm

* run cargo fmt

* lint

* resolve all conflicts

* update release note and add more tests

* call_limit does not need to be explicitly set to None

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>

* simplify iteration in gc protocol

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>

* update vf2 doc

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>

* cargo fmt + remove unnecessary mut ref in semantic matcher

* more tests

* fix clippy error

* fix failing test case

* Fix docstring typo

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>