Rework definition to treat disconnected nodes as 0
This commit updates the definition of the average path length for
disconnected nodes to use the definition from [1] where the path is
treated as 0 from source to target if there is no path. This removes
the need to check connectivity up front (which was being done
incorrectly for directed graphs) and lets us remove the is_connected()
function added in previous commits.

[1] https://en.wikipedia.org/wiki/Average_path_length#Definition