Add unweighted average shortest path length function (#403)
* Add unweighted average shortest path length function

This commit adds a new function unweighted_average_shortest_path_length.
This mirrors the default behvaior of networkx's
average_shortest_path_length() function without any weight function or
attribute specified. When we add support for weighted methods (using
dijkstra and floyd warshall) we can add a similar unified function like
networkx's.

* Add missing parameter to docstrings

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

* Simplify neighbors iteration logic

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

* Handle graph edge cases

This commit adds handling for 3 edge cases with the input graph, it's
disconnected, it's empty, or it's a single node and returns an
appropriate early return (NaN, 0, inf respectively). To handle the
disconnected case a new helper function is_connected() is added to the
connectivity module which does a quick bfs traversal from a node in the
graph to see if all other nodes are reachable or not.

* Fix python lint

* Reverse node count casting

* Rework definition to treat disconnected nodes as 0

This commit updates the definition of the average path length for
disconnected nodes to use the definition from [1] where the path is
treated as 0 from source to target if there is no path. This removes
the need to check connectivity up front (which was being done
incorrectly for directed graphs) and lets us remove the is_connected()
function added in previous commits.

[1] https://en.wikipedia.org/wiki/Average_path_length#Definition

* Remove incorrect docstring about disconnected graphs

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

* Apply suggestions from code review

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

* Add as_undirected tests

* Add missing as_undirected flag to negative tests