Add Heavy hex code to generators (#293)
Added implementation of heavy hex [1], the input is the code distance, d. In the implementation I chose to add nodes referring to each type of qubit used in the Heavy Hex Code (data, syndrome, and flag) which may cause the node order to be confusing, but that was the best way that I figure out how to implement. Now I will explain the way that I implemented the code, which could be better for people without experience with the heavy hex code.

In the heavy hex code, there are d ** 2 data qubits, (d + 1) * (d - 1) / 2 syndrome qubits, and d * (d - 1) flag qubits. In order to implement I follow 3 steps:

 * Each flag qubit is conneceted with 2 data qubits forming a chain on the graph, the chain has size d.
 * Connect syndrome qubits that are connected to data qubits.
 * Connect flag and syndrome qubits.

In steps (ii) and (iii), there are two patterns depending on the eve (odd) columns of syndrome qubits, which are treated accordingly using an if statement. It was also required to use .chunk() for Vectors in order to represent them in a column.

Related to #150

[1] https://journals.aps.org/prx/abstract/10.1103/PhysRevX.10.011022

* Added heavy hex

Added firs implementation for heavy hex, it only implements 1 1 heavy hex and has some hardcoded parts that will be worked on.

* Added to docs

* lint

* lint

* add heavy hex generator

* lint

* add release notes

* add test edge assertion

* fix merge

* add doc references

* upgrade legibilitty

* fix doc identation

* add code-block for ASCII

* add ascii block

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>

* add `directed_heavy_hex_graph`

* updated release notes

* Fix edge directions

* Fix docstring typos

Co-authored-by: Eric Peterson <peterson.eric.c@gmail.com>

* Updates tests with direction fixes

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>
Co-authored-by: Eric Peterson <peterson.eric.c@gmail.com>