Prepare 0.10.0 release (#414)
* Prepare 0.10.0 release

This commit starts the preparation for the 0.10.0 release. It moves all
the release notes for the release into a separate folder (to distinguish
them from notes for future releases) and also reworks the content of the
release notes and updates the documentation for the release. This commit
should be the last one merged before releasing 0.10.0 to ensure we've
moved and updated all the release notes before the release.

* Update releasenotes/notes/0.10.0/prepare-0.10.0-b8c2251507f0d13d.yaml

Co-authored-by: Corey Mendell <corey.mendell@gmail.com>

* Fix whitespace issues

* Move and update new release notes

* Move and update new releasenote

* Fix betweenness_centrality docstring sphinx syntax

Co-authored-by: Corey Mendell <corey.mendell@gmail.com>