Workaround is_isormorphic returning false for 2 empty graphs
This commit works around the bug in the vf2 implementation for empty
graphs where no mapping is found for is_isomorphic. It's just a
workaround for the issue reported in #421 and I'm sure there is a better
real fix, but in the interest of having a solution proposed in the short
term this was fast.

Fixes #421