Fix VF2 in case of empty graphs.
An alternative for #421. This commit fixes the output of functions
solving the isomorphism problem (`is_isomorphic`, `is_subgraph_isomorphic`,
`vf2_mapping`) without handling empty graphs as a special case but rather
they now fall naturally into the logic of the algorithm. The solution is to
equivalently check if we have found a complete mapping in the `Outer` phase
and not in the `Inner` phase of the algorithm, where we dont reach in case
of an empty graph.