Parallelize shortest path calculation in betweenness_centrality()
This commit updates the internals of the betweenness_centrality()
function to leverage rayon parallel iterators to parallelize the loop
over all nodes in the graph which finds the shortest path for
centrality. There was no data dependency for doing this calculation over
all nodes and it is only using the return that does. This changes the
implementation to pre-compute the list of ShortestPathData objects first
in parallel and then iterate over that, instead of iterating over the
nodes and computing it as we go. The tradeoff here is that we'll use
more memory to store the intermediate results in exchange for being able
to execute in parallel. Benchmarking will need to be done to see if this
is worthwhile and if there is a threshold where the tradeoff makes
sense.