Eliminate intermediate storage and parallelize up to rescale
This commit reworks the parallelization to run the full betweenness
calculation loop in parallel. This wasn't done in the initial commit
because the betweeness Vec is mutated by each thread and we can't write
to it from multiple threads at once. To overcome this a RwLock is used
to ensure only one thread is writing to the betweenness Vec at once.
While this does limit the overall throughput of the function  it's not
any slower than doing it serially and it eliminates the use of
intermediate storage between the main loop and updating the output
betweeness Vec.