Reduce lock wait time
The previous commit which introduced the use of a RwLock around the
betweenness Vec was a bit to aggressive in which code required the lock.
It was aquiring a lock too early which resulted in each thread having a
long wait while it waited for the lock, despite there being work it
could do before it needed the lock. We only need the lock when
betweenness is being used in the parallel portion of the code. This
commit updates the use of the lock to do just this which further
improves the runtime of the function.

Before adopting this approach I did try to go a step further than this
commit and acquire the lock right only for each individual write to
betweenness, but this significantly degraded performance. The overhead
of locking in that case became quite high and at least from a quick
inspection each thread was spending ~50% of it's time dealing with
locking. The approach used in this commit resulted in much better
performance.