Pin nightly version in coverage job (#433)
The coverage ci job uses the rust nightly channel to enable the coverage
collection. However, building with nightly has recently stopped working
because of a change breaking num-bigint (see
https://github.com/rust-num/num-bigint/issues/218 for more details).
Until the rustc nightly behavior is reverted or a new num-bigint version
is released that is compatible with the new nightly behavior this commit
pins the nightly version used in the coverage job to a version prior to
the failure.