Make from_adjacency_matrix generic and add PyGraph version
This commit updates the implementation of the from_adjacency_matrix()
functions to have a shared generic private function to be shared across
the per type variants. To accomplish this the IsNan trait from rust
nightly is copied into retworkx (so we can build with stable rust). If
in the future if an int type variant is added we'll need to implement
this trait for the int type (and just return false).

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>