Prepare 0.10.2 release
This commit prepares for a 0.10.2 bugfix release to get the recent vf2
bugfixes on main out in the wild. Looking at the git log since 0.10.1
we've only merged fixes and requirements bumps since the release, so now
is a good time to push out the bugfix release before we start merging
new features.