Improve betweenness centrality release note
Reviewing the release notes for this pending release I noticed that the
release note for the betweenness centrality function was a bit lacking.
This commit updates it to include more details and example using and
visualizing it. Reno lets us edit existing release notes like this
because it associates the file creation in the git history with the tag
so subsequent edits of existing note files are still under the correct
release.