Prepare 0.10.2 release (#434)
* Prepare 0.10.2 release

This commit prepares for a 0.10.2 bugfix release to get the recent vf2
bugfixes on main out in the wild. Looking at the git log since 0.10.1
we've only merged fixes and requirements bumps since the release, so now
is a good time to push out the bugfix release before we start merging
new features.

* Update deps

* Fix release note syntax

* Update releasenotes/notes/0.10.2/prepare-0.10.2-d75b2d86be66a6d9.yaml

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

* Move new release note

* Minor rewording in prelude

* Improve betweenness centrality release note

Reviewing the release notes for this pending release I noticed that the
release note for the betweenness centrality function was a bit lacking.
This commit updates it to include more details and example using and
visualizing it. Reno lets us edit existing release notes like this
because it associates the file creation in the git history with the tag
so subsequent edits of existing note files are still under the correct
release.

* Fix release note typo

* Tweak betweenness centrality release note a bit more

* Another typo fix

* Move new release notes