Bump version strings post release (#442)
Now that the 0.10.2 release is out the door we should bump the version
string on main to show the source version we're installing is newer
than what's been released. This bumps the version strings to 0.11.0. If
we decide to do another bugfix release in the 0.10.x series we can
create a branch from 0.10.2 and go from there.