Parallelize betweenness_centrality() (#428)
* Parallelize shortest path calculation in betweenness_centrality()

This commit updates the internals of the betweenness_centrality()
function to leverage rayon parallel iterators to parallelize the loop
over all nodes in the graph which finds the shortest path for
centrality. There was no data dependency for doing this calculation over
all nodes and it is only using the return that does. This changes the
implementation to pre-compute the list of ShortestPathData objects first
in parallel and then iterate over that, instead of iterating over the
nodes and computing it as we go. The tradeoff here is that we'll use
more memory to store the intermediate results in exchange for being able
to execute in parallel. Benchmarking will need to be done to see if this
is worthwhile and if there is a threshold where the tradeoff makes
sense.

* Eliminate intermediate storage and parallelize up to rescale

This commit reworks the parallelization to run the full betweenness
calculation loop in parallel. This wasn't done in the initial commit
because the betweeness Vec is mutated by each thread and we can't write
to it from multiple threads at once. To overcome this a RwLock is used
to ensure only one thread is writing to the betweenness Vec at once.
While this does limit the overall throughput of the function  it's not
any slower than doing it serially and it eliminates the use of
intermediate storage between the main loop and updating the output
betweeness Vec.

* Reduce lock wait time

The previous commit which introduced the use of a RwLock around the
betweenness Vec was a bit to aggressive in which code required the lock.
It was aquiring a lock too early which resulted in each thread having a
long wait while it waited for the lock, despite there being work it
could do before it needed the lock. We only need the lock when
betweenness is being used in the parallel portion of the code. This
commit updates the use of the lock to do just this which further
improves the runtime of the function.

Before adopting this approach I did try to go a step further than this
commit and acquire the lock right only for each individual write to
betweenness, but this significantly degraded performance. The overhead
of locking in that case became quite high and at least from a quick
inspection each thread was spending ~50% of it's time dealing with
locking. The approach used in this commit resulted in much better
performance.

* Remove leftover lines from abandoned approach

* Add parallel_threshold kwarg

This commit adds a new kwarg parallel_threshold which is used to set the
number of nodes that the function starts to run multithreaded. For any
graph with less nodes than the value of this kwarg the execution will be
single threaded.

* Run cargo fmt

* Fix python docstring lint

* Add parallel test coverage