Add from complex matrix function (#412)
* Add from complex matrix function

This commit adds a new constructor method from_complex_adjacency_matrix
that will build a new graph from an adjacency matrix with complex
elements.

Fixes #411

* Make from_adjacency_matrix generic and add PyGraph version

This commit updates the implementation of the from_adjacency_matrix()
functions to have a shared generic private function to be shared across
the per type variants. To accomplish this the IsNan trait from rust
nightly is copied into retworkx (so we can build with stable rust). If
in the future if an int type variant is added we'll need to implement
this trait for the int type (and just return false).

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

* Add tests

* Add release notes

* Fix copy paste issues in do strange

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

* Remove unused Zero trait bound on generic function

* Iterate over enumerate elements instead of index range

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>