Add support and CI config for Python 3.10
The release of Python 3.10 is coming soon on: 10/04/2021. As of this
commit the second release candidate has been published which should be
stable to start testing. This commit adds the CI test jobs for all
platforms with Python 3.10 and updates the release wheel jobs to
including building python 3.10 wheels. In the interest of CI time this
commit also drops the win32 platform from the test matrix, as it
provides limited coverage over the win64 jobs and would be 5 (including
3.10) extra jobs.

Fixes #415