Fix `distance_matrix` and `k_shortest_path_lengths` if node holes.
Previously, `distance_matrix` and `k_shortest_path` would panic
if the input graph contained holes in node indices. This commit
fixes this in `distance_matrix` by mapping nodes to a compact interval
`0..n` where `n` is the number of nodes in the graph.
In `k_shortest_path` we should initialize a vector of `node_bound` size.