Handle d=1 edge case in heavy graph functions
The *heavy_hex_graph and *heavy_square_graph generator functions all
previously had an issue where they would panic when d=1 was passed as
the input to the functions. The d=1 is just a single node, so to avoid a
panic a special case is added to return a graph with a single node which
avoids the panic.

Fixes #452